internal/lsp: refactor workspace Symbol method
There are a number of issues with the current implementation:

* test variant packages are not handled correctly, meaning duplicate
  symbols are returned
* fuzzy results are not ordered by score

We refactor the implementation of workspace symbol to use a
symbolCollector that carries context during the walk for symbols. As
part of resolving the test variant issue, we first determine a list of
packages to walk.

(*symbolCollector).viewsToPackages gathers the packages we are going to
inspect for symbols.  This pre-step is required in order to filter out
any "duplicate" *types.Package. The duplicates arise for packages that
have test variants.  For example, if package mod.com/p has test files,
then we will visit two packages that have the PkgPath() mod.com/p: the
first is the actual package mod.com/p, the second is a special version
that includes the non-XTest _test.go files. If we were to walk both of
of these packages, then we would get duplicate matching symbols and we
would waste effort. Therefore where test variants exist we walk those
(because they include any symbols defined in non-XTest _test.go files)

One further complication is that even after this filtering, packages
between views might not be "identical" because they can be built using
different build constraints (via the "env" config option).

Therefore on a per view basis we first build up a map of PkgPath() ->
*types.Package preferring the test variants if they exist. Then we merge
the results between views, de-duping by *types.Package.

Finally, when we come to walk these packages and start gathering
symbols, we ignore any symbols we have already seen (different
*types.Package for the same import path because of different build
constraints), keeping track of those symbols via symbolCollector.seen.

Then we walk that list of packages in much the same way as before.

To fix the issue of fuzzy resuts not being ordered by score, we define a
symbolMatcher interface type, along with two implementations: fuzzy and
case-based. It is the responsibility of a symbolMatcher to gather
matches and then present results.

Error handling is simplified through the use of the
(*symbolCollector).check method, and accompany defer-recover check in
(*symbolCollector).walk. At any point the symbol walk may be ended early
using the sentienl error earlySymbolWalkExit.

Refactoring also makes it easier to make later changes for advanced
query syntax.

Change-Id: I6d130f0bac31cc42e45eb1b15cdb2a53cdca4ae4